Update and rename 3b1b Essence of LA.pdf to 3b1b Essence of Linear Algebra.pdf
These are notes from the very popular 3b1b Essence of Linear Algebra YouTube playlist.  I hope these notes will be helpful for refreshing the concepts or a quick revision before interviews. That said, I would highly recommend to watch the videos and use these notes only as a companion, not a replacement.